refactor(auth): move page-boot helpers to www/dashboard
build_auth_context / _onboarding_complete / _provider_logins / _provider_is_configured are dashboard page-load concerns, not an HTTP API (none is whitelisted), and only www/dashboard.get_context used them. Move them beside get_context; api/auth keeps only the signup endpoints. test_auth (6) green.
